requirements.txt
spacy==3.2.0
transformers==4.16.2
pandas
scikit-learn
packaging==20.0
pytorch-lightning==1.5.10
tqdm==4.41.1
hydra-core==1.1.1
clearml==1.2.0
typing-extensions==3.10.0.2
focal_loss_torch==0.0.9
